rename: 管理しているフォルダの名前を変更
