+++ lib/main.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:git_repository/presentation/app.dart';
+import 'package:git_repository/app/app.dart';
 
 void main() {
   runApp(const MainApp());
